+++ example/lib/main.dart
@@ -25,6 +25,7 @@ class _MyAppState extends State<MyApp> {
 
     polar = Polar();
     polar.heartRateStream.listen((e) => log('Heart rate: ${e.data.hr}'));
+    polar.batteryLevelStream.listen((e) => log('Battery: ${e.level}'));
     polar.streamingFeaturesReadyStream.listen((e) {
       if (e.features.contains(DeviceStreamingFeature.ecg)) {
         polar